# contracts/openapi/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Event-Driven E-commerce Identity API
  version: 1.0.0
  description: Authentication, sessions, roles, service principals, and security operations for the platform.
servers:
  - url: https://identity.ecommerce.local
    description: Local and portfolio deployment placeholder
tags:
  - name: Authentication
  - name: Sessions
  - name: Users
  - name: Keys
  - name: Service identity
  - name: Operations
paths:
  /health/live:
    get:
      tags: [Operations]
      summary: Check process liveness
      operationId: healthLive
      security: []
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
  /health/ready:
    get:
      tags: [Operations]
      summary: Check dependency readiness
      operationId: healthReady
      security: []
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /health/startup:
    get:
      tags: [Operations]
      summary: Check startup completion
      operationId: healthStartup
      security: []
      responses:
        '200':
          description: Startup completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /metrics:
    get:
      tags: [Operations]
      summary: Expose Prometheus metrics
      operationId: metrics
      security: []
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain:
              schema: { type: string }
  /_test/mailbox/latest:
    get:
      tags: [Operations]
      summary: Read the latest local test email token
      description: Local development only. This route is registered only when IDENTITY_TEST_MAILER_ENABLED=true and must never be enabled in production.
      operationId: getLatestTestMailboxMessage
      x-local-only: true
      security: []
      parameters:
        - name: kind
          in: query
          required: true
          schema: { type: string, enum: [email-verification, password-reset] }
        - name: recipient
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        '200':
          description: Latest one-time token captured by the local simulated mailer
          content:
            application/json:
              schema:
                type: object
                required: [kind, recipient, token]
                properties:
                  kind: { type: string }
                  recipient: { type: string, format: email }
                  token: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
  /.well-known/register:
    post:
      tags: [Authentication]
      summary: Register a demo user
      description: Development/demo only. Disabled unless demo registration is explicitly enabled.
      operationId: registerDemoUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate a human user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Rotated session tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current session
      operationId: logout
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Session revoked or already revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/logout-all:
    post:
      tags: [Authentication]
      summary: Revoke all sessions for the current user
      operationId: logoutAll
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: All sessions revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/sessions:
    get:
      tags: [Sessions]
      summary: List the current user's sessions
      operationId: listSessions
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Active and recent sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/password-reset/request:
    post:
      tags: [Authentication]
      summary: Request a password reset
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { description: Accepted without revealing account existence }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/password-reset/confirm:
    post:
      tags: [Authentication]
      summary: Confirm a password reset
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string, minLength: 32 }
                new_password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
      responses:
        '204': { description: Password reset and sessions revoked }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/email-verification/confirm:
    post:
      tags: [Authentication]
      summary: Confirm an email verification token
      operationId: confirmEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 32 }
      responses:
        '204': { description: Email verified }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/auth/service-token:
    post:
      tags: [Service identity]
      summary: Issue a machine access token
      description: Internal-only client-credentials endpoint. Never expose through the public gateway.
      operationId: issueServiceToken
      x-internal: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceTokenRequest' }
      responses:
        '200':
          description: Machine access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceTokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /.well-known/jwks.json:
    get:
      tags: [Keys]
      summary: Get public signing keys
      operationId: getJwks
      responses:
        '200':
          description: JSON Web Key Set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JsonWebKeySet' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/users/{user_id}:
    get:
      tags: [Users]
      summary: Get a user visible to the caller
      operationId: getUser
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/users/{user_id}/roles:
    post:
      tags: [Users]
      summary: Assign a role
      operationId: assignRole
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, pattern: '^[a-z][a-z0-9-]{1,31}$' }
      responses:
        '204': { description: Role assigned }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/users/{user_id}/roles/{role}:
    delete:
      tags: [Users]
      summary: Remove a role
      operationId: removeRole
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: role
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z][a-z0-9-]{1,31}$' }
      responses:
        '204': { description: Role removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/users/{user_id}/status:
    patch:
      tags: [Users]
      summary: Change a user's status
      operationId: changeUserStatus
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserStatusRequest' }
      responses:
        '204': { description: User status changed and affected sessions revoked when required }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema: { type: string, pattern: '^usr_[A-Za-z0-9]+$' }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Unauthorized:
      description: Authentication failed
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Forbidden:
      description: Authenticated but not permitted
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Conflict:
      description: Resource conflict
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    TooManyRequests:
      description: Rate limit exceeded
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    ServiceUnavailable:
      description: Required dependency unavailable
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    RefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token: { type: string, minLength: 32, writeOnly: true }
    ServiceTokenRequest:
      type: object
      required: [client_id, client_secret, scopes]
      additionalProperties: false
      properties:
        client_id: { type: string, minLength: 3, maxLength: 128 }
        client_secret: { type: string, minLength: 32, writeOnly: true }
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 128 }
    ServiceTokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      additionalProperties: false
      properties:
        access_token: { type: string, readOnly: true }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, const: 300, minimum: 1 }
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string, readOnly: true }
        refresh_token: { type: string, readOnly: true }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, const: 600, minimum: 1 }
        refresh_expires_in: { type: integer, minimum: 1, default: 2592000 }
    HealthResponse:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { type: string, enum: [ok, ready, starting] }
    User:
      type: object
      required: [id, email, status, roles, created_at]
      properties:
        id: { type: string, pattern: '^usr_[A-Za-z0-9]+$' }
        email: { type: string, format: email }
        status: { type: string, enum: [ACTIVE, DISABLED, PENDING_VERIFICATION] }
        roles: { type: array, items: { type: string } }
        email_verified_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    UserStatusRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { type: string, enum: [ACTIVE, DISABLED] }
    Session:
      type: object
      required: [id, family_id, status, created_at, expires_at]
      properties:
        id: { type: string }
        family_id: { type: string }
        status: { type: string, enum: [ACTIVE, ROTATED, REVOKED] }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
    JsonWebKeySet:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            type: object
            required: [kty, crv, kid, use, alg, x]
            properties:
              kty: { type: string, const: OKP }
              crv: { type: string, const: Ed25519 }
              kid: { type: string }
              use: { type: string, const: sig }
              alg: { type: string, const: EdDSA }
              x: { type: string, description: Base64url-encoded public key }
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        request_id: { type: string }